states added, constants changed, functions added to drivecontrol, collect item order changed for 1 and 5

+++ src/supergoodrobot/parameters.h
@@ -11,8 +11,8 @@ namespace parameters {
     extern uint8_t basespeed;
     extern uint8_t qrdthreshold;
 	//extern uint8_t irthreshold;
-	extern uint8_t sideqrdthreshold;
-	extern uint8_t sideirthreshold;
+	extern uint16_t sideqrdthreshold;
+	extern uint16_t sideirthreshold;
 }
 
 #endif
